+++ lib/lastfm_archive/utils.ex
@@ -2,11 +2,12 @@ defmodule LastfmArchive.Utils do
   @moduledoc false
 
   alias LastfmArchive.Behaviour.Archive
+  require Logger
 
-  @data_dir Application.compile_env(:lastfm_archive, :data_dir, "./archive_data/")
+  @data_dir Application.compile_env(:lastfm_archive, :data_dir, "./lastfm_data/")
   @metadata_file ".archive"
 
-  @file_io Application.compile_env(:lastfm_archive, :file_io)
+  @file_io Application.compile_env(:lastfm_archive, :file_io, Elixir.File)
 
   @doc """
   Generate {from, to} daily time ranges for querying Last.fm API based on
@@ -35,40 +36,28 @@ defmodule LastfmArchive.Utils do
     {DateTime.to_unix(from), DateTime.to_unix(to)}
   end
 
-  def year_range({from, to}) do
-    DateTime.from_unix!(from).year..DateTime.from_unix!(to).year
-  end
+  def year_range({from, to}), do: DateTime.from_unix!(from).year..DateTime.from_unix!(to).year
 
   def data_dir(options \\ []), do: Keyword.get(options, :data_dir, @data_dir)
   def user_dir(user, options \\ []), do: Path.join([data_dir(options), user])
+
   def metadata_filepath(user, options), do: Path.join([data_dir(options), user, @metadata_file])
+  def num_pages(playcount, per_page), do: (playcount / per_page) |> :math.ceil() |> round
 
-  def display_progress(archive) do
-    IO.puts("Archiving #{archive.extent} scrobbles for #{archive.creator}")
+  # returns 2021/12/31/200_001 type paths
+  def page_path(datetime, page, per_page) do
+    page_num = page |> to_string() |> String.pad_leading(3, "0")
+
+    datetime
+    |> DateTime.from_unix!()
+    |> DateTime.to_date()
+    |> Date.to_string()
+    |> String.replace("-", "/")
+    |> Path.join("#{per_page}_#{page_num}")
   end
 
-  def display_progress({from, _to}, playcount, pages) do
-    from_date = DateTime.from_unix!(from) |> DateTime.to_date()
-
-    IO.puts("\n")
-    IO.puts("#{from_date}")
-    IO.puts("#{playcount} scrobble(s)")
-    IO.puts("#{pages} page(s)")
-  end
-
-  def display_skip_message({from, _to}, playcount) do
-    from_date = DateTime.from_unix!(from) |> DateTime.to_date()
-
-    IO.puts("\n")
-    IO.puts("Skipping #{from_date}, previously synced: #{playcount} scrobble(s)")
-  end
-
-  def display_api_error_message({from, _to}, reason) do
-    from_date = DateTime.from_unix!(from) |> DateTime.to_date()
-
-    IO.puts("\n")
-    IO.puts("Last.fm API error while syncing #{from_date}: #{reason}")
-  end
+  def date(from) when is_integer(from), do: DateTime.from_unix!(from) |> Calendar.strftime("%Y-%m-%d")
+  def date({from, _day}) when is_integer(from), do: DateTime.from_unix!(from) |> Calendar.strftime("%Y-%m-%d")
 
   @doc """
   Read and unzip a file from the archive of a Lastfm user.
@@ -95,5 +84,31 @@ defmodule LastfmArchive.Utils do
     dir = Path.join(user_dir(user, []), "tsv")
     unless @file_io.exists?(dir), do: @file_io.mkdir_p(dir)
     :ok
+  end
+
+  @doc """
+  Write scrobbles (map) data to a file in the archive of a Lastfm user.
+  """
+  def write(archive, scrobbles, options \\ [])
+
+  def write(%Archive{creator: creator}, scrobbles, options) when is_map(scrobbles) do
+    with metadata_filepath <- metadata_filepath(creator, options),
+         path <- get_filepath(options) do
+      full_path =
+        metadata_filepath
+        |> Path.dirname()
+        |> Path.join("#{path}.gz")
+
+      full_path_dir = Path.dirname(full_path)
+      unless @file_io.exists?(full_path_dir), do: @file_io.mkdir_p(full_path_dir)
+      @file_io.write(full_path, scrobbles |> Jason.encode!(), [:compressed])
+    end
+  end
+
+  def write(_archive, {:error, api_message}, _options), do: {:error, api_message}
+
+  defp get_filepath(options) do
+    path = Keyword.get(options, :filepath)
+    if path != nil and path != "", do: path, else: raise("please provide a valid :filepath option")
   end
 end